fix: refactor file paths
