+++ python/dify_plugin/entities/tool.py
@@ -17,55 +17,7 @@ from dify_plugin.core.utils.yaml_loader import load_yaml_file
 from dify_plugin.entities import I18nObject
 from dify_plugin.entities.model.message import PromptMessageTool
 from dify_plugin.entities.oauth import OAuthSchema
-
-
-class LogMetadata(str, Enum):
-    STARTED_AT = "started_at"
-    FINISHED_AT = "finished_at"
-    ELAPSED_TIME = "elapsed_time"
-    TOTAL_PRICE = "total_price"
-    TOTAL_TOKENS = "total_tokens"
-    PROVIDER = "provider"
-    CURRENCY = "currency"
-
-
-class CommonParameterType(Enum):
-    SECRET_INPUT = "secret-input"
-    TEXT_INPUT = "text-input"
-    SELECT = "select"
-    STRING = "string"
-    NUMBER = "number"
-    FILE = "file"
-    FILES = "files"
-    BOOLEAN = "boolean"
-    APP_SELECTOR = "app-selector"
-    MODEL_SELECTOR = "model-selector"
-    # TOOL_SELECTOR = "tool-selector"
-    TOOLS_SELECTOR = "array[tools]"
-
-
-class AppSelectorScope(Enum):
-    ALL = "all"
-    CHAT = "chat"
-    WORKFLOW = "workflow"
-    COMPLETION = "completion"
-
-
-class ModelConfigScope(Enum):
-    LLM = "llm"
-    TEXT_EMBEDDING = "text-embedding"
-    RERANK = "rerank"
-    TTS = "tts"
-    SPEECH2TEXT = "speech2text"
-    MODERATION = "moderation"
-    VISION = "vision"
-
-
-class ToolSelectorScope(Enum):
-    ALL = "all"
-    PLUGIN = "plugin"
-    API = "api"
-    WORKFLOW = "workflow"
+from dify_plugin.entities.provider_config import CommonParameterType, LogMetadata, ProviderConfig
 
 
 class ToolRuntime(BaseModel):
@@ -281,42 +233,6 @@ class ToolLabelEnum(Enum):
 class ToolCredentialsOption(BaseModel):
     value: str = Field(..., description="The value of the option")
     label: I18nObject = Field(..., description="The label of the option")
-
-
-class ProviderConfig(BaseModel):
-    class Config(Enum):
-        SECRET_INPUT = CommonParameterType.SECRET_INPUT.value
-        TEXT_INPUT = CommonParameterType.TEXT_INPUT.value
-        SELECT = CommonParameterType.SELECT.value
-        BOOLEAN = CommonParameterType.BOOLEAN.value
-        MODEL_SELECTOR = CommonParameterType.MODEL_SELECTOR.value
-        APP_SELECTOR = CommonParameterType.APP_SELECTOR.value
-        # TOOL_SELECTOR = CommonParameterType.TOOL_SELECTOR.value
-        TOOLS_SELECTOR = CommonParameterType.TOOLS_SELECTOR.value
-
-        @classmethod
-        def value_of(cls, value: str) -> "ProviderConfig.Config":
-            """
-            Get value of given mode.
-
-            :param value: mode value
-            :return: mode
-            """
-            for mode in cls:
-                if mode.value == value:
-                    return mode
-            raise ValueError(f"invalid mode value {value}")
-
-    name: str = Field(..., description="The name of the credentials")
-    type: Config = Field(..., description="The type of the credentials")
-    scope: str | None = None
-    required: bool = False
-    default: Optional[Union[int, float, str]] = None
-    options: Optional[list[ToolCredentialsOption]] = None
-    label: I18nObject
-    help: Optional[I18nObject] = None
-    url: Optional[str] = None
-    placeholder: Optional[I18nObject] = None
 
 
 class ToolProviderIdentity(BaseModel):
